requirements.txt
# Core dependencies
pywebview>=4.0

# For packaging (optional)
pyinstaller>=5.0